[PRDP-64] feat: Updated chart, refactored package, introducing agents download inside Docker file, updated values
