+++ agent/config.yaml
@@ -1,0 +1,2 @@
+rules:
+  - pattern: ".*"
