+++ src/main/java/it/gov/pagopa/pdf/engine/Info.java
@@ -1,0 +1,30 @@
+package it.gov.pagopa.pdf.engine;
+
+import com.microsoft.azure.functions.*;
+import com.microsoft.azure.functions.annotation.AuthorizationLevel;
+import com.microsoft.azure.functions.annotation.FunctionName;
+import com.microsoft.azure.functions.annotation.HttpTrigger;
+
+import java.util.Optional;
+
+/**
+ * Azure Functions with Azure Http trigger.
+ */
+public class Info {
+
+	/**
+	 * This function will be invoked when a Http Trigger occurs
+	 * @return {@link HttpResponseMessage} with {@link HttpStatus#OK}
+	 */
+	@FunctionName("Info")
+	public HttpResponseMessage run (
+			@HttpTrigger(name = "InfoTrigger",
+			methods = {HttpMethod.GET},
+			route = "info",
+			authLevel = AuthorizationLevel.ANONYMOUS) HttpRequestMessage<Optional<String>> request,
+			final ExecutionContext context) {
+
+		return request.createResponseBuilder(HttpStatus.OK).build();
+	}
+
+}
