+++ src/main/java/it/gov/pagopa/pdf/engine/exception/GeneratePDFException.java
@@ -1,0 +1,46 @@
+/*
+Copyright (C)
+
+This program is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public
+License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later
+version.
+
+This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
+details.
+
+You should have received a copy of the GNU Affero General Public License along with this program.
+If not, see https://www.gnu.org/licenses/.
+*/
+
+package it.gov.pagopa.pdf.engine.exception;
+
+import it.gov.pagopa.pdf.engine.model.AppErrorCodeEnum;
+
+/**
+ * Thrown in case of problems when generating the PDFA/2a document
+ */
+public class GeneratePDFException extends PDFEngineException {
+
+    /**
+     * Constructs new exception with provided error code and message
+     *
+     * @param errorCode Error code
+     * @param message Detail message
+     */
+    public GeneratePDFException(AppErrorCodeEnum errorCode, String message) {
+        super(errorCode, message);
+    }
+
+    /**
+     * Constructs new exception with provided error code, message and cause
+     *
+     * @param errorCode Error code
+     * @param message Detail message
+     * @param cause Exception causing the constructed one
+     */
+    public GeneratePDFException(AppErrorCodeEnum errorCode, String message, Throwable cause) {
+        super(errorCode, message, cause);
+    }
+
+}
