+++ src/main/java/it/gov/pagopa/pdf/engine/util/Constants.java
@@ -1,0 +1,9 @@
+package it.gov.pagopa.pdf.engine.util;
+
+public class Constants {
+
+    public static final String UNZIPPED_FILES_FOLDER = "/unzipped";
+    public static final String ZIP_FILE_NAME = "/input.zip";
+
+    private Constants() {}
+}
