--- src/main/java/it/gov/pagopa/project/util/Constants.java
@@ -1,9 +1,0 @@
-package it.gov.pagopa.project.util;
-
-public class Constants {
-
-    public static final String UNZIPPED_FILES_FOLDER = "/unzipped";
-    public static final String ZIP_FILE_NAME = "/input.zip";
-
-    private Constants() {}
-}
