--- src/test/java/it/gov/pagopa/project/InfoTest.java
@@ -1,47 +1,0 @@
-package it.gov.pagopa.project;
-
-import com.microsoft.azure.functions.ExecutionContext;
-import com.microsoft.azure.functions.HttpRequestMessage;
-import com.microsoft.azure.functions.HttpResponseMessage;
-import com.microsoft.azure.functions.HttpStatus;
-import it.gov.pagopa.project.util.HttpResponseMessageMock;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.Spy;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.mockito.stubbing.Answer;
-
-import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
-
-@ExtendWith(MockitoExtension.class)
-class InfoTest {
-
-	@Mock
-    ExecutionContext executionContextMock;
-
-    @Spy
-    Info sut;
-
-    @Test
-    void runOK() {
-        @SuppressWarnings("unchecked")
-        HttpRequestMessage<Optional<String>> request = mock(HttpRequestMessage.class);
-
-        doAnswer((Answer<HttpResponseMessage.Builder>) invocation -> {
-            HttpStatus status = (HttpStatus) invocation.getArguments()[0];
-            return new HttpResponseMessageMock.HttpResponseMessageBuilderMock().status(status);
-        }).when(request).createResponseBuilder(any(HttpStatus.class));
-
-        // test execution
-        HttpResponseMessage response = sut.run(request, executionContextMock);
-
-        // test assertion
-        assertEquals(HttpStatus.OK, response.getStatus());
-    }
-
-}
